Breaking up util: files.go functions moved
ExpandGlobs moved to repos.go: It's only used there.

FilterPaths moved to git/util.go: It's only used in the git pkg.

Both functions are now private since they're never used outside their
respective packages.
