--- util/files.go
@@ -1,93 +1,0 @@
-package util
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-
-	"github.com/G-Node/gin-cli/ginclient/log"
-)
-
-// PathSplit returns the directory path separated from the filename. If the
-// argument is a directory, the filename "." is returned. Errors are ignored.
-func PathSplit(path string) (string, string) {
-	info, err := os.Stat(path)
-	if err != nil {
-		return path, "."
-	}
-	var dir, filename string
-	switch mode := info.Mode(); {
-	case mode.IsRegular():
-		dir, filename = filepath.Split(path)
-	case mode.IsDir():
-		dir = path
-		filename = "."
-	}
-	return dir, filename
-}
-
-// ExpandGlobs expands a list of globs into paths (files and directories).
-// An error is returned if at least one element of the input slice does not match a real path.
-func ExpandGlobs(paths []string) (globexppaths []string, err error) {
-	if len(paths) == 0 {
-		// Nothing to do
-		globexppaths = paths
-		return
-	}
-	// expand potential globs
-	for _, p := range paths {
-		log.LogWrite("ExpandGlobs: Checking for glob expansion for %s", p)
-		exp, globerr := filepath.Glob(p)
-		if globerr != nil {
-			log.LogWrite(globerr.Error())
-			log.LogWrite("Bad file pattern %s", p)
-			return nil, globerr
-		}
-		if exp == nil {
-			log.LogWrite("ExpandGlobs: No files matched")
-			return nil, fmt.Errorf("No files matched %v", p)
-		}
-		globexppaths = append(globexppaths, exp...)
-	}
-	if len(globexppaths) == 0 {
-		// Invalid paths
-		log.LogWrite("ExpandGlobs: No files matched")
-		err = fmt.Errorf("No files matched %v", paths)
-	}
-	return
-}
-
-func stringInSlice(element string, strlist []string) bool {
-	for _, str := range strlist {
-		if element == str {
-			return true
-		}
-	}
-	return false
-}
-
-// FilterPaths takes path descriptions (full paths, relative paths, files, or directories) and returns a slice of filepaths (as strings) excluding the files listed in excludes.
-func FilterPaths(paths, excludes []string) (filtered []string) {
-
-	// walker adds files to filtered if they don't match excludes
-	walker := func(path string, info os.FileInfo, err error) error {
-		if filepath.Base(path) == ".git" {
-			return filepath.SkipDir
-		}
-		if info.IsDir() {
-			return nil
-		}
-		if !stringInSlice(path, excludes) {
-			filtered = append(filtered, path)
-		}
-		return nil
-	}
-
-	for _, p := range paths {
-		err := filepath.Walk(p, walker)
-		if err != nil {
-			log.LogWrite("Error occured during path filtering: %s", err.Error())
-		}
-	}
-	return
-}
